+++ presentation-core/src/main/java/tachiyomi/presentation/core/components/AdaptiveSheet.kt
@@ -1,11 +1,10 @@
 package tachiyomi.presentation.core.components
 
-import android.content.res.Configuration.ORIENTATION_LANDSCAPE
 import androidx.activity.compose.BackHandler
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
-import androidx.compose.animation.rememberSplineBasedDecay
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.gestures.AnchoredDraggableDefaults
 import androidx.compose.foundation.gestures.AnchoredDraggableState
 import androidx.compose.foundation.gestures.DraggableAnchors
 import androidx.compose.foundation.gestures.Orientation
@@ -24,10 +23,12 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
@@ -38,7 +39,6 @@ import androidx.compose.ui.input.nestedscroll.NestedScrollConnection
 import androidx.compose.ui.input.nestedscroll.NestedScrollSource
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.layout.onSizeChanged
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.Velocity
@@ -49,24 +49,16 @@ import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.launch
 import kotlin.math.roundToInt
 
-private val sheetAnimationSpec = tween<Float>(durationMillis = 350)
-
 @Composable
 fun AdaptiveSheet(
     isTabletUi: Boolean,
-    enableSwipeDismiss: Boolean,
+    enableImplicitDismiss: Boolean,
     onDismissRequest: () -> Unit,
     modifier: Modifier = Modifier,
     content: @Composable () -> Unit,
 ) {
     val density = LocalDensity.current
     val scope = rememberCoroutineScope()
-    val maxWidth = if (LocalConfiguration.current.orientation == ORIENTATION_LANDSCAPE) {
-        600.dp
-    } else {
-        460.dp
-    }
-
     if (isTabletUi) {
         var targetAlpha by remember { mutableFloatStateOf(0f) }
         val alpha by animateFloatAsState(
@@ -85,6 +77,7 @@ fun AdaptiveSheet(
                 .clickable(
                     interactionSource = null,
                     indication = null,
+                    enabled = enableImplicitDismiss,
                     onClick = internalOnDismissRequest,
                 )
                 .fillMaxSize()
@@ -93,7 +86,7 @@ fun AdaptiveSheet(
         ) {
             Surface(
                 modifier = Modifier
-                    .requiredWidthIn(max = maxWidth)
+                    .requiredWidthIn(max = 460.dp)
                     .clickable(
                         interactionSource = null,
                         indication = null,
@@ -105,7 +98,10 @@ fun AdaptiveSheet(
                 shape = MaterialTheme.shapes.extraLarge,
                 color = MaterialTheme.colorScheme.surfaceContainerHigh,
                 content = {
-                    BackHandler(enabled = alpha > 0f, onBack = internalOnDismissRequest)
+                    BackHandler(
+                        enabled = enableImplicitDismiss && remember { derivedStateOf { alpha > 0f } }.value,
+                        onBack = internalOnDismissRequest,
+                    )
                     content()
                 },
             )
@@ -115,16 +111,14 @@ fun AdaptiveSheet(
             }
         }
     } else {
-        val decayAnimationSpec = rememberSplineBasedDecay<Float>()
-        val anchoredDraggableState = remember {
-            AnchoredDraggableState(
-                initialValue = 1,
-                positionalThreshold = { with(density) { 56.dp.toPx() } },
-                velocityThreshold = { with(density) { 125.dp.toPx() } },
-                snapAnimationSpec = sheetAnimationSpec,
-                decayAnimationSpec = decayAnimationSpec,
-            )
+        val anchoredDraggableState = rememberSaveable(saver = AnchoredDraggableState.Saver()) {
+            AnchoredDraggableState(initialValue = 1)
         }
+        val flingBehavior = AnchoredDraggableDefaults.flingBehavior(
+            state = anchoredDraggableState,
+            positionalThreshold = { _: Float -> with(density) { 56.dp.toPx() } },
+            animationSpec = sheetAnimationSpec,
+        )
         val internalOnDismissRequest = {
             if (anchoredDraggableState.settledValue == 0) {
                 scope.launch { anchoredDraggableState.animateTo(1) }
@@ -149,19 +143,19 @@ fun AdaptiveSheet(
         ) {
             Surface(
                 modifier = Modifier
-                    .widthIn(max = maxWidth)
+                    .widthIn(max = 460.dp)
                     .clickable(
                         interactionSource = null,
                         indication = null,
                         onClick = {},
                     )
                     .then(
-                        if (enableSwipeDismiss) {
+                        if (enableImplicitDismiss) {
                             Modifier.nestedScroll(
                                 remember(anchoredDraggableState) {
-                                    anchoredDraggableState.preUpPostDownNestedScrollConnection(
-                                        onFling = { scope.launch { anchoredDraggableState.settle(it) } },
-                                    )
+                                    anchoredDraggableState.preUpPostDownNestedScrollConnection {
+                                        scope.launch { anchoredDraggableState.settle(sheetAnimationSpec) }
+                                    }
                                 },
                             )
                         } else {
@@ -181,7 +175,8 @@ fun AdaptiveSheet(
                     .anchoredDraggable(
                         state = anchoredDraggableState,
                         orientation = Orientation.Vertical,
-                        enabled = enableSwipeDismiss,
+                        enabled = enableImplicitDismiss,
+                        flingBehavior = flingBehavior,
                     )
                     .navigationBarsPadding()
                     .statusBarsPadding(),
@@ -189,7 +184,7 @@ fun AdaptiveSheet(
                 color = MaterialTheme.colorScheme.surfaceContainerHigh,
                 content = {
                     BackHandler(
-                        enabled = anchoredDraggableState.targetValue == 0,
+                        enabled = enableImplicitDismiss && anchoredDraggableState.targetValue == 0,
                         onBack = internalOnDismissRequest,
                     )
                     content()
@@ -235,7 +230,7 @@ private fun <T> AnchoredDraggableState<T>.preUpPostDownNestedScrollConnection(
 
     override suspend fun onPreFling(available: Velocity): Velocity {
         val toFling = available.toFloat()
-        return if (toFling < 0 && offset > 0f) {
+        return if (toFling < 0 && offset > anchors.minPosition()) {
             onFling(toFling)
             // since we go to the anchor with tween settling, consume all for the best UX
             available
@@ -246,7 +241,11 @@ private fun <T> AnchoredDraggableState<T>.preUpPostDownNestedScrollConnection(
 
     override suspend fun onPostFling(consumed: Velocity, available: Velocity): Velocity {
         onFling(available.toFloat())
-        return available
+        return if (targetValue != settledValue) {
+            available
+        } else {
+            Velocity.Zero
+        }
     }
 
     private fun Float.toOffset(): Offset = Offset(0f, this)
@@ -257,3 +256,5 @@ private fun <T> AnchoredDraggableState<T>.preUpPostDownNestedScrollConnection(
     @JvmName("offsetToFloat")
     private fun Offset.toFloat(): Float = this.y
 }
+
+private val sheetAnimationSpec = tween<Float>(durationMillis = 350)